refactor: align adapter directory naming
