+++ host/services/supervisor/internal/adapters/bus_inproc/bus.go
